Refactor VAT calculations and update test operations
Eliminated explicit casting and made computations more straightforward in 'CalculateVAT' function. These changes provide clearer and more manageable code that is easier to understand. Changes in the test file 'expensesClassificationDoc_test.go' reflects corresponding modifications in the function. 'math' package was imported in 'enums.go' to easily round down the computed VAT to cents.

+++ mydata/mydataInvoices/expensesClassificationDoc_test.go
@@ -137,7 +137,7 @@ func (suite *ExpenseClassificationSuite) TestInvoiceExpensesClassificationNewWay
 	// vat amount testing
 	require.Equal(t, (*float64)(nil), inv1Entries[0].VatAmount)
 	require.Equal(t, (*float64)(nil), inv1Entries[1].VatAmount)
-	require.Equal(t, float64(28.8), *inv1Entries[2].VatAmount)
+	require.Equal(t, 28.8, *inv1Entries[2].VatAmount)
 	require.Equal(t, float64(0), *inv1Entries[3].VatAmount)
 
 	// classification category
